backend/requirements.txt
weaviate-client
python-dotenv
openai
black
wasabi
typer
farm-haystack
nltk
fastapi
uvicorn